chore: Fix spec
